Split LSP code-action handlers into server_codeaction.go
Move handleCodeAction, the quick-fix / source.fixAll.mdsmith builders,
the previewFix capability gate, and the WorkspaceEdit construction
helpers (fullFileEdit, fullFileEditAnnotated, annotatedHunkEdits,
documentEndPosition) out of server.go. Pure file-split refactor — no
behavior change. server.go: 1362 -> 935 lines (now under the 1,000-line
threshold).
